refactoring and bug fixes
